feat: Faz 1 butunluk — para decimal, seans guard'lari, cocuk tek-acik-seans
- Company fiyatlari int->decimal; EF HasPrecision(12,2) tum para alanlarina
- BookingService.Create: cocuk zorunlu + cocuk-basina tek acik seans
  (Sistem Misafiri 999999 muaf) + unique-violation yakalama
- BookingService.Update: tamamlama kilidi (Completed yalniz close akisi),
  bitmis-seans dokunulmazligi, gecis matrisi
- scripts/faz1-booking-child-guard.sql: kismi unique index
  ux_one_open_booking_per_child + company kolon numeric garantisi (DB'ye uygulandi)
- tests: 9 yeni BookingServiceTests; MapperTests decimal'e uyarlandi

+++ Models/DTOs/Company/CompanyDto.cs
@@ -6,8 +6,8 @@ namespace Iroh.Models.DTOs.Company
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public int FirstHourPrice { get; set; }
-        public int AdditionalHalfHourPrice { get; set; }
+        public decimal FirstHourPrice { get; set; }
+        public decimal AdditionalHalfHourPrice { get; set; }
 
         public static CompanyDto From(Models.Entities.Company c) => new()
         {
